Replaced exceptions with std::expected in the argument parser

+++ src/datetime.cpp
@@ -45,8 +45,8 @@ uint64_t DateTimeToTimestamp(const DateTime& datetime) {
     return result;
 }
 
-int16_t ParseIntValue(std::string_view str) {
-    int16_t result;
+int64_t ParseIntValue(std::string_view str) {
+    int64_t result;
     std::from_chars_result convertion_result = std::from_chars(str.data(), str.data() + str.size(), result);
 
     if (convertion_result.ec == std::errc::invalid_argument || convertion_result.ptr != str.end()) {
